+++ doc/src/changelog.md
@@ -4,6 +4,7 @@ Version History
 ### New Features in Embree 3.5.0
 -   Changed normal oriented curve definition to fix waving artefacts.
 -   Fixed bounding issue for normal oriented motion blurred curves.
+-   Fixed performance issue with motion blurred point geometry.
 
 ### New Features in Embree 3.4.0
 -   Added point primitives (spheres, ray-oriented discs, normal-oriented discs).